[Bug 277] here are the patches that Mandrake and Debian use Merged in relevant vendor fixes for 64-bit compatibility. Thanks Lee and Patrice for sifting through the detritus!

+++ faxd/PCFFont.c++
@@ -233,7 +233,7 @@ PCFFont::read(const char* name)
 	}
 	for (i = 0; i < nbitmaps; i++) {
 	    metrics[i].bits = bitmaps + offsets[i];
-	    if ((unsigned) metrics[i].bits & 1) {
+	    if ((unsigned long) metrics[i].bits & 1) {
 		error("Internal error, bitmap data not word-aligned");
 		delete offsets;
 		return (false);
